requirements.txt
asgiref==3.3.1
astroid==2.5.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-jsonfield==1.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==2.1.0
pylint==2.7.2
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
